Move `HaloRadar` to a separate file
